+++ test/ints.rs
@@ -2,34 +2,34 @@
 #![allow(dead_code, unused_attributes)]
 
 #[miri_run]
-fn ret() -> i32 {
+fn ret() -> i64 {
     1
 }
 
 #[miri_run]
-fn neg() -> i32 {
+fn neg() -> i64 {
     -1
 }
 
 #[miri_run]
-fn add() -> i32 {
+fn add() -> i64 {
     1 + 2
 }
 
 #[miri_run]
-fn indirect_add() -> i32 {
+fn indirect_add() -> i64 {
     let x = 1;
     let y = 2;
     x + y
 }
 
 #[miri_run]
-fn arith() -> i32 {
+fn arith() -> i64 {
     3*3 + 4*4
 }
 
 #[miri_run]
-fn match_int() -> i32 {
+fn match_int() -> i64 {
     let n = 2;
     match n {
         0 => 0,
@@ -41,7 +41,7 @@ fn match_int() -> i32 {
 }
 
 // #[miri_run(expected = "Int(4)")]
-// fn match_int_range() -> i32 {
+// fn match_int_range() -> i64 {
 //     let n = 42;
 //     match n {
 //         0...9 => 0,
